# Fix deadlock when stopping a shard