If a queue is stopped and one of its shards happens to hit the
batch_send_deadline at the same time a deadlock can occur where stop
holds the mutex and will not release it until the send is finished, but
the send needs the mutex to retrieve the most recent batch. This is
fixed by using a second mutex just for writing.

In addition, the test I wrote exposed a case where during shutdown a
batch could be sent twice due to concurrent calls to queue.Batch() and
queue.FlushAndShutdown(). Protect these with a mutex as well.

Fixes #10264 

This change is a bit slower due to the extra mutexes, but still much faster
than before #9830. Some work could be done in the future to remove some
mutexes/speed this up but this is the safest change to fix the issue I came
up with.
```
name           old time/op    new time/op    delta
SampleSend-16    1.64ms ± 2%    1.75ms ± 1%  +6.79%  (p=0.000 n=10+9)

name           old alloc/op   new alloc/op   delta
SampleSend-16    50.9kB ± 8%    51.9kB ± 5%    ~     (p=0.247 n=10+10)

name           old allocs/op  new allocs/op  delta
SampleSend-16       148 ± 1%       150 ± 1%  +1.35%  (p=0.001 n=9+9)
```

cc: @beorn7 this is pointing to the release-2.33 branch.